+++ src/pages/index-controller/components/components/big-button/big-button.component.js
@@ -1,0 +1,5 @@
+module.exports = {
+    handleClick(event, selected) {
+        this.emit('click', selected);
+    }
+}